+++ Server/app/index.js
@@ -1,19 +1,17 @@
 const express = require('express');
 const app = express();
-const router = require("./routes")
+const routeHandler = require("./routes");
 
+// Middleware to parse JSON bodies
 app.use(express.json());
-// localhost:3000/
-app.get('/', (req, res) => {
+
+app.get("/", (req, res) => {
     res.status(200).json({
-        message: "GET - root",
-        metadata: {
-            hostname: req.hostname, 
-            method: req.method, 
-        },
+        message: "API in running", 
+        success: true
     });
 });
 
-app.use('/api', router);
+app.use("/api/v1", routeHandler);
 
 module.exports = app;