+++ Server/server.js
@@ -1,9 +1,8 @@
-const http = require('http');
-require('dotenv').config();
-const app = require('./app');
+require("dotenv").config();
+const app = require("./app");
 
-const server = http.createServer(app);
+const PORT = process.env.PORT || 3000;
 
-server.listen(process.env.PORT, () => {
-    console.log(`Server is running on port ${process.env.PORT}`);
-});
+app.listen(PORT, () => {
+    console.log(`Server is running on port ${PORT}`);
+})